+++ src/Phan/Output/Collector/BufferingCollector.php
@@ -56,7 +56,7 @@ final class BufferingCollector implements IssueCollectorInterface
             return;
         }
         if (self::$trace_issues) {
-            CLI::printToStderr("Backtrace of $issue is:\n");
+            CLI::printToStderr("Backtrace of {$issue->getIssue()->getType()} $issue is:\n");
             if (self::$trace_issues === Issue::TRACE_VERBOSE) {
                 \phan_print_backtrace();
             } else {